extract grid functions from date adapter
