+++ packages/Schedulely/src/dateAdapters/date.ts
@@ -103,15 +103,6 @@ export const createDefaultAdapter = (locale = 'en'): DateTimeAdapter => {
   const addMonthsToDate = (date: Date, amount: number) =>
     new Date(date.getFullYear(), date.getMonth() + amount, 1);
 
-  const getGridStartIndex = (eventDate: Date, startOfWeek: Date) =>
-    eventDate <= startOfWeek ? 1 : eventDate.getDay() + 1; //add one because css-grid isn't zero-index'd
-
-  const getGridEndIndex = (eventEndDate: Date, endOfWeek: Date) => {
-    if (eventEndDate > endOfWeek) return 8;
-    const end = eventEndDate.getDay() + 2; // add two because css-grid isn't zero index'd, and day of week is zero-index'd
-    return end;
-  };
-
   const isEventInWeek = (
     eventStartDate: Date,
     eventEndDate: Date,
@@ -154,8 +145,6 @@ export const createDefaultAdapter = (locale = 'en'): DateTimeAdapter => {
     isSameMonth,
     isDateToday,
     addMonthsToDate,
-    getGridEndIndex,
-    getGridStartIndex,
     isEventInWeek,
     convertIsoToDate,
     isCurrentMonth,
