+++ packages/Schedulely/src/layouts/eventWeekLayout/EventWeekLayout.tsx
@@ -1,7 +1,6 @@
 import { InternalCalendarEvent } from '@/types/InternalCalendarEvent';
 import {
   useActions,
-  useCalendar,
   useComponents,
   useEventHighlight,
   useEventIntersection,
@@ -12,14 +11,20 @@ interface EventLayoutProps {
   daysInweek: Date[];
 }
 
+const getGridStartIndex = (eventDate: Date, startOfWeek: Date) =>
+  eventDate <= startOfWeek ? 1 : eventDate.getDay() + 1; //add one because css-grid isn't zero-index'd
+
+const getGridEndIndex = (eventEndDate: Date, endOfWeek: Date) => {
+  if (eventEndDate > endOfWeek) return 8;
+  const end = eventEndDate.getDay() + 2; // add two because css-grid isn't zero index'd, and day of week is zero-index'd
+  return end;
+};
+
 /**
  * This component controls the layout of an individual events within a week
  * @returns EventLayout Component
  */
 export const EventWeekLayout = ({ events, daysInweek }: EventLayoutProps) => {
-  const {
-    dateAdapter: { getGridStartIndex, getGridEndIndex },
-  } = useCalendar();
   const { eventComponent: EventComponent } = useComponents();
   const { setHighlight, clearHighlight, isHighlighted } = useEventHighlight();
   const { onEventClick } = useActions();
